+++ test/integration/helpers.py
@@ -21,8 +21,16 @@ BUCKET_NAME_LENGTH = 50
 BUCKET_CREATED_AT_MILLIS = 'created_at_millis'
 
 
-def bucket_name_part(length):
+def _bucket_name_prefix_part(length: int) -> str:
     return ''.join(random.choice(BUCKET_NAME_CHARS) for _ in range(length))
+
+
+def get_bucket_name_prefix(rnd_len: int = 8) -> str:
+    return GENERAL_BUCKET_NAME_PREFIX + _bucket_name_prefix_part(rnd_len)
+
+
+def random_bucket_name(prefix: str = GENERAL_BUCKET_NAME_PREFIX) -> str:
+    return prefix + _bucket_name_prefix_part(BUCKET_NAME_LENGTH - len(prefix))
 
 
 def authorize(b2_auth_data, api_config=DEFAULT_HTTP_API_CONFIG):